refactor search page to suspense client component
